Add SHA updating logic for remotes

and additional changes from the current version of devtools

+++ R/install-remote.R
@@ -8,8 +8,31 @@
 #'   \item calls install
 #' }
 #' @noRd
-install_remote <- function(remote, ..., quiet = FALSE) {
+install_remote <- function(remote, ..., force = FALSE, quiet = FALSE) {
   stopifnot(is.remote(remote))
+
+  remote_sha <- remote_sha(remote)
+  package_name <- remote_package_name(remote)
+  local_sha <- local_sha(package_name)
+
+  if (!isTRUE(force) &&
+    !different_sha(remote_sha = remote_sha, local_sha = local_sha)) {
+
+    if (!quiet) {
+      message(
+        "Skipping install of '", package_name, "' from a ", sub("_remote", "", class(remote)[1L]), " remote,",
+        " the SHA1 (", substr(remote_sha, 1L, 8L), ") has not changed since last install.\n",
+        "  Use `force = TRUE` to force installation")
+    }
+    return(invisible(FALSE))
+  }
+
+  if (inherits(remote, "cran_remote")) {
+    install_packages(
+      package_name, repos = remote$repos, type = remote$pkg_type,
+      ..., quiet = quiet)
+    return(invisible(TRUE))
+  }
 
   bundle <- remote_download(remote, quiet = quiet)
   on.exit(unlink(bundle), add = TRUE)
@@ -58,3 +81,134 @@ is.remote <- function(x) inherits(x, "remote")
 
 remote_download <- function(x, quiet = FALSE) UseMethod("remote_download")
 remote_metadata <- function(x, bundle = NULL, source = NULL) UseMethod("remote_metadata")
+remote_package_name <- function(remote, ...) UseMethod("remote_package_name")
+remote_sha <- function(remote, ...) UseMethod("remote_sha")
+
+remote_package_name.default <- function(remote, ...) remote$repo
+remote_sha.default <- function(remote, ...) NA_character_
+
+different_sha <- function(remote_sha = NULL,
+                          local_sha = NULL) {
+  if (is.null(remote_sha)) {
+    remote_sha <- remote_sha(remote)
+  }
+
+  if (is.null(local_sha)) {
+    local_sha <- local_sha(remote_package_name(remote))
+  }
+
+  same <- remote_sha == local_sha
+  same <- isTRUE(same) && !is.na(same)
+  !same
+}
+
+local_sha <- function(name) {
+  if (!is_installed(name)) {
+    return(NA_character_)
+  }
+  package2remote(name)$sha %||% NA_character_
+}
+
+# Convert an installed package to its equivalent remote. This constructs the
+# remote from metadata stored in the package's DESCRIPTION file; the metadata
+# is added to the package when it is installed by remotes. If the package is
+# installed some other way, such as by `install.packages()` there will be no
+# meta-data, so there we construct a generic CRAN remote.
+package2remote <- function(name, lib = .libPaths(), repos = getOption("repos"), type = getOption("pkgType")) {
+
+  x <- tryCatch(utils::packageDescription(name, lib.loc = lib), error = function(e) NA, warning = function(e) NA)
+
+  # will be NA if not installed
+  if (identical(x, NA)) {
+    return(remote("cran",
+        name = name,
+        repos = repos,
+        pkg_type = type,
+        sha = NA_character_))
+  }
+
+  if (is.null(x$RemoteType)) {
+
+    # Packages installed with install.packages() or locally without remotes
+    return(remote("cran",
+        name = x$Package,
+        repos = repos,
+        pkg_type = type,
+        sha = x$Version))
+  }
+
+  switch(x$RemoteType,
+    github = remote("github",
+      host = x$RemoteHost,
+      repo = x$RemoteRepo,
+      subdir = x$RemoteSubdir,
+      username = x$RemoteUsername,
+      ref = x$RemoteRef,
+      sha = x$RemoteSha),
+    gitlab = remote("gitlab",
+      host = x$RemoteHost,
+      repo = x$RemoteRepo,
+      subdir = x$RemoteSubdir,
+      username = x$RemoteUsername,
+      ref = x$RemoteRef,
+      sha = x$RemoteSha),
+    xgit = remote("xgit",
+      url = x$RemoteUrl,
+      branch = x$RemoteBranch,
+      sha = x$RemoteSha,
+      subdir = x$RemoteSubdir,
+      args = x$RemoteArgs),
+    git2r = remote("git2r",
+      url = x$RemoteUrl,
+      branch = x$RemoteBranch,
+      sha = x$RemoteSha,
+      subdir = x$RemoteSubdir),
+    bitbucket = remote("bitbucket",
+      host = x$RemoteHost,
+      repo = x$RemoteRepo,
+      username = x$RemoteUsername,
+      ref = x$RemoteRef,
+      sha = x$RemoteSha,
+      subdir = x$RemoteSubdir),
+    svn = remote("svn",
+      url = x$RemoteUrl,
+      svn_subdir = x$RemoteSubdir,
+      revision = x$RemoteSha,
+      args = x$RemoteArgs),
+    local = remote("local",
+      path = x$RemoteUrl,
+      subdir = x$RemoteSubdir,
+      sha = {
+        # Packages installed locally might have RemoteSha == NA_character_
+        x$RemoteSha %||% x$Version
+      }),
+    url = remote("url",
+      url = x$RemoteUrl,
+      subdir = x$RemoteSubdir,
+      config = x$RemoteConfig,
+      pkg_type = x$RemotePkgType),
+    bioc_git2r = remote("bioc_git2r",
+      mirror = x$RemoteMirror,
+      repo = x$RemoteRepo,
+      release = x$RemoteRelease,
+      sha = x$RemoteSha,
+      branch = x$RemoteBranch),
+    bioc_xgit = remote("bioc_xgit",
+      mirror = x$RemoteMirror,
+      repo = x$RemoteRepo,
+      release = x$RemoteRelease,
+      sha = x$RemoteSha,
+      branch = x$RemoteBranch),
+
+    # packages installed with install_cran
+    cran = remote("cran",
+      name = x$Package,
+      repos = eval(parse(text = x$RemoteRepos)),
+      pkg_type = x$RemotePkgType,
+      sha = x$RemoteSha))
+}
+
+#' @export
+format.remotes <- function(x, ...) {
+  vapply(x, format, character(1))
+}
